Fix Choices.__getitem__ access for Python 3.
For Python 3, __getitem__ fails because filter returns an iterable for
`filter` instead of a list. Because it's an iterable, accessing a
Choice class like `ChoiceObj['MY_CHOICE']` fails with:

TypeError: 'filter' object is not subscriptable

This branch provides an alternate solution that is functionally the
same without using `filter`. Tests are included to improve coverage.